requirements.txt
fastapi
python-jose
passlib
mysql-connector
mysql-connector-python
python-multipart
websockets
uvicorn